util refactor, cutsec changed place, bi-lisa mention removed from nmr
